Verify auth before deletion or update of UserSettings [Processing]

+++ src/CarCare.Processing/Services/UserSettingsService.cs
@@ -1,6 +1,7 @@
 ﻿using CarCare.Persistence.Interfaces;
 using CarCare.Persistence.Models;
 using CarCare.Processing.Interfaces.Service;
+using CarCare.Processing.Interfaces.Session;
 using CarCare.Processing.Models.Core;
 
 namespace CarCare.Processing.Services;
@@ -8,10 +9,12 @@ namespace CarCare.Processing.Services;
 internal class UserSettingsService : IUserSettingsService
 {
     private readonly IUserSettingsRepository _settingsRepository;
+    private readonly IUserSession _session;
 
-    public UserSettingsService(IUserSettingsRepository settingsRepository)
+    public UserSettingsService(IUserSettingsRepository settingsRepository, IUserSession session)
     {
         _settingsRepository = settingsRepository;
+        _session = session;
     }
 
     public async Task AddAsync(UserSettings settings)
@@ -22,6 +25,11 @@ internal class UserSettingsService : IUserSettingsService
 
     public async Task DeleteAsync(Guid userId)
     {
+        if (!_session.IsAuthenticated(userId))
+        {
+            return;
+        }
+
         await _settingsRepository.DeleteAsync(userId);
     }
 
@@ -34,6 +42,11 @@ internal class UserSettingsService : IUserSettingsService
 
     public async Task UpdateAsync(UserSettings settings)
     {
+        if (!_session.IsAuthenticated(settings.UserId))
+        {
+            return;
+        }
+
         UserSettingsDao dao = settings.ToDao();
         await _settingsRepository.UpdateAsync(dao);
     }
